[Anti-Cheat] Further changes to mob agro detection
Prevention of players being able to flash log to loose mob agro.  All
AIs should now be brough up when a user logs in.  Changed from prior
method of fooling the AI into thinking we moved, now we count logging in
as a move.  This can be further enhanced by writing flash loggers to
cheater logs by counting the number of logins in a short period of time
and the delta of the time they were offline however the cheaters log
portions remain to be written