[FIX] Bond
Several fixes to make Bond work as I believe it was intended.  The spell
currently works as described here:

* Bond had been changed at some point to stop working as it originally
did and was simply hacked to make an item transcedent (soft white light,
stays with you through death).  Being that you could find these types
items fairly easily by killing mobs and the benefit wasn't all that
great by itself, this spell was basically broken in my opinion.  It took
so long and so many regs to cast for the benefit it would never be used.
The item attribute Bonded is still present for things like Riija swords.

1.  Create new item attribute for SoulBind (iasbin.kod) to differentiate
Bond from transcendent.  This is so you can continue to have items drop
off mobs in one way, and have Bond end up with a different result.  Also
this new attribute can be leveraged later for unique items and or vanity
items given out by the gods, won at tournaments or the like but still be
different than Bonded or Transcendent items so attributes can be set
uniquely.
2.  Fix casting time so it is not listed twice (Template problem again,
now uses actual casting time)
3.  Fix further template problems with spell power usage so it is not
using the default and can be adjusted in the spell itself rather than
having to adjust a global
4.  Fix for items sometimes not being able to be picked up (I.E. Error
message: You cannot pickup the scimitar) after the item was bonded.
5.  Attributes for DropOnDeath, Mending, Brittle, Enchant, Shatter,
Swap, Identify and Spell Modifier defined in the iasbind.kod for easy
game balancing if needed.
6.  Added proper message to Bond description and weapon description,
"Infuses an item with a heavenly oratorio", to differentiate from
transcendent weapons.
7.  Changed the number of heartstones required to four (overkill?) so
you really have to "want" to Bond that MSH to protect it from baddies
but then you can actually use it outside an inn :)

Bonded items now:
1. Stay with you upon death
2. Can be mended (default behavior of most items)
3. Cannot be brittled
4. Can be enchanted (per the normal restrictions on enchanting
items/weapons)
5. Cannot be swapped
6. Cannot be shattered
7. Can be identified (default behavior of most items)

* It might be a thought to make bonded items unique, untradable, and or
undroppable.  Due to the heavy reagent charge this should not be a
problem, but a change can be made to limit the number of soulbound
items.  Possibly raising or lowering the number of heartstones by itself
may be an option.

As always please feel free to try the changes and report any undocument
features or bugs to me.