Mapupate: Undergroundlake
This commit add a secondary entrace for brax. It should come together
with a hometownpen, where also user under 30 hps getting ported to the
hometown after a pen. This update makes brax-mules not longer nassary.
It needs a ironkey in the underground lake to step into brax. KOD by
Taran.