Update kod/object/active/holder/nomoveon/battler/player.kod
I went ahead to made a change suggested by Gar to increase the caps to be at the same levels as monsters. This shouldn't mean that players are now as powerful as boss mobs as the real challenge with boss monsters is the amount of damage they do combined with the skills they use.

Again this is to increase the weight of aim and agility to be closer to being as important as any other stat at end game.