Attempt to make Eviltwin useful
It attacks very fast.
It dies in 1 or 2 hits.
There is a limit of one per room and I actually think that's the best
way to limit it. Sort of like earthquake. If another team wants to get a
twin up all they have to do is kill the current twin and take over.

I tested it a lot now and I think this is the sweet spot. It takes
awhile for it to kill a built player and that player can easily turn
around and mange it, but I feel as if it's a neat little distraction
that adds a bit of damage. The proc rate can be nutty because of how
fast it attacks but because it's hitting you with your weapon I feel
like that's on you.