Monster Add: Beetle
It's not direct a push to the source commit - its more like, that i need
help from someone on creating the bbg file. Zaphod want wo have
everything buildable from the source. But the UI of BBGun is a pain
(cause its old and there is not any documentation) and i dont understand
the content of the bbg-file right. So it would be nice when a codejunky
would go into it.

This new mob based on the Ant and would work if you replace simple the
ant.bgf in your client to the bgf file included in the Zip-File i hosted
on (www.mark-sturm.de/dl/m59/beetle/beetle.zip).

The grey beelte is strached on a factor of 6 and the redbeetle of a
factor of 4.