Jig change
Jig (the other Truce-like spell) also had a No_Combat flag attached to
it. I just removed that aspect of the spell. It still makes players
dance, and it still prevents equipping of items, it just doesn't stop
combat anymore (which was the problematic abuse we've been seeing).

I think there are some rather interesting applications for preventing
gear switching (which the spell still does) but not enough for it to be
worth abusing from a mule.