Loyalty Potion + Signet Ring Quest Mod
This idea was dreamed up by Mark.

*NOTE: Fixed makefile to no longer be dependant on other changes (i.e.
Nkarmpot/Pkarmapot) per Andrew's Request.

1) Created Loyalty Potion:

Will increase HP by 1 if player is < 40 HP
If player is > 40 HP, or Outlaw, or Murderer then it will just delete
itself and not increase HP.
2) Changed Signet Ring quest to reward this Potion.

Both Mark and I believe that this is a trivial change that will make the
signet ring quest useful again. Signet rings are not created often
enough to farm and make an army of 40 HP characters. This will be a fun
dynamic for newbies and oldies to work and bring the signet ring back to
"life" per se.