Changed weight to 1,000
Changed weight to 1,000

A person with 1 might can pickup 1 potion.
A person with 50 might can pickup 2 potions.
A person with 50 might + SS can pickup 3 potions.

All of this is ignored when the potion is given to the player as a
reward regardless of what they are holding because it is given with the
NewHold message putting it directly into their inventory regardless of
if they are full weight or not.

This removes any ability for people to farm them at high rates, it just
becomes impractical - however they still hold their unique and decent
value to keep the quest somewhat useful.