Destroy Walls at entrances
Added a new function to rooms that is called by players entering an
area. It destroys walls in a small radius around the player, eliminating
the 'trick' aspect of wall spells.

This setup allows walls to be used on exits to block escapes, but does
not allow walls to block entrances. Walls will still function normally
anywhere else in the area, and can still block small passages and the
like, but players will no longer 'stumble' into them.