Removes Guardian Angel after 2 months
This removes all of the mules deaths, all of the logoff ghosts of mules,
and basically makes mules a little more dangerous. Maybe people will
utilize mules less.

When a player is born they get a flag that says "This soul is new to the
lands of Meridian 59."

That lasts for 2 months (2 meridian years). Once that 2 months is over
that is removed from their bio and they lose their guardian angel.

Anyone can get above 30 hps in 2 months.