Nerudite bow tweak
Nerudite bow had a spellpower penalty of 0, rather than -30 like battle
bows. That really didn't make sense (nerudite is supposed to be
antimagic), and it was also incredibly powerful for veterans.

This puts nerudite bows at -30 power penalty, but increases their chance
to shatter nerudite arrows, which is what everyone always thought they
did before we got a look at the code. This change equates to 30% higher
chance to add 1-2 damage. Shouldn't break anything, but is not at all
shabby.