Remove Guardian Angel after 2 months. V2
Automatically checks each time a character logs on to see if they are
over 2 months old, if they are sends message PkillEnable, which removes
guardian angel, sends Vulnerable message and sends "Congratulations"
game mail describing that they no longer need their Guardian Angel.

Notes:

When a player is born they get an honor string that says "This soul is
new to the
lands of Meridian 59."

That lasts for 2 months (2 meridian years). Once that 2 months is over
that is removed from their bio and now they lose their guardian angel.

As well as other benefits...

This removes the ability for mules to be used as scouts
Removes the ability for mules to just die to move around the world
Makes leaving mules around more dangerous
Makes mules required to logoff in safe areas....
** Compiles without errors, no server errors, tested thoroughly.