Comments & Battler
Changed the function to DeleteWallsAroundBattler. Given that it works
for monsters, too, it could definitely be used by content creators to
make monsters that inherently clear walls. Really, the applications are
up to whoever calls the function. I just happen to also put in a call
that occurs every time a player enters an area. I updated the comments
to reflect that.