MSH Adjustments
I think we can agree MSH's for how rare they are -- are essentially a
useless item used for looks. I would like to increase the attributes of
an MSH from a slight increase from a regular helmet to that of a much
better HELM.

* Decreased durability slightly.

* Decreased spell modifier (reducing power of spells cast by helm's
user) since it is "anti-magic".

* Increased Defense Base to act more like a good piece of armor -
increases dodge chance.

* Doubled it's attack spell reduction from 15 to 30 since it is
considered an Anti Magic piece of Armor.

This should make the risk of losing or using an MSH worth while.

One thing I am not 100% on, existing helms - will this change their
makeup?? Or does somehting special have to happen?