Summoned Object list
Created a list of a player's summoned objects. Moved Reflection's
delete-upon-logoff code to the player's logoffhook as an example of how
the list can be used (instead of the reflections constantly checking
their caster's online status, we just delete them directly when the
player logs off).

No actual in-game changes included.