RoomGive admin command
Gives all non-DM players in the room a reward, similar to GlobalGive.

who = a user. That user's room will be given the items. Admin should
usually call it on himself. (It's easier to look up player #s than it is
room #s).

hp = min basemaxhealth required to receive the items. Default is 30.