ShadowRift Portal timer to close
Added a timer to the constructor to close the portal after a given time.

Made a timer tied to viMultiCast_SpellPower (which is used to determine
how long it takes to cast the spell) which is also editable in the $
screen in game. My thought process behind this was, making the spell
harder to cast should/could make the portal last longer.

Necro Portals will now last a random time between the lowest timer is 20
mins and the highest timer is 30 mins after timer elapses the portal
closes itself.

viMultiCast_SpellPower is currently 1500
1500 * 800 = 20 mins
1500 * 1200 = 30 mins

I think 20-30 mins is fair.