Guildhall & Portal Tweaks
- Flipping switch no longer turns player outlaw
- Conquering no longer turns player outlaw
- Allow unguildeds and immature guild members to flip switch
- Allow portals in halls. Initially this was because angeled characters
can now be killed in halls, but:
- Prisms cannot be bound to halls, so players can't portal people into
their hall to kill them free of consequence.

Altogether this should eliminate all of the problematic safety and check
issues guildhalls have, while allowing players to open portals in them
(just not *to* them). This is all relevant because I'd like to add new
things to guildhalls, especially mobility/portal options (since the
alternate rescue location is such a big hit), and these old awkward
mechanics needed to be fixed up to make that possible.