Statue Annoyance Fix
Per much player request, changed living statues' wake up chance to 100%
instead of 70%.

Seriously, why would they do this to us? Waiting for statues to wake up
is the worst. Fail a few rolls and you can be sitting idle on the same
statue for 10 minutes or more. I see no reason to keep normal statues
this way, but if admins want to do something for an event with it, the
code is still there.