Wall Limit to Antistack change
This is step one of retooling wall spells into far more active and
interesting spells now that the 'zone in' problem, that had been acting
as the primary restraint on wall power, has been dealt with.

This commit changes the summoning limit in a room to 200, up from 18.
Instead, rooms now have a new anti-stacking function which prevents more
than one element of the same type to be placed on the same spot. This
should prove to be a far more organic limit on wall spell use, along
with vigor cost, mana cost, and duration rebalancing.