Blind & Dazzle fix
There is a widespread cheat to ignore the visual component of these
flagship spells. This commit therefore largely eliminates those effects,
reducing the visual components to one second (retained as a way to
inform players they have been cast upon).

This necessitated a few text changes in dazzle, explaining the dazzle
effect as a pacification effect.

Blind was more complicated. I renamed it "Bind" and increased the
stumbling effect from 30% to 70% to compensate for the loss of blinding.
This should affect the movement of normal players and cheaters alike in
the manner blind was originally intended to.