Set & Loadout - robes, scrolls
Added a special case for Robes of the Disciple, which were complicated
to handle. Robes now have a GetSchool function so other objects can tell
what school they are.

Removed Set & Loadout as default spells. Instead, all innkeepers now
sell Set & Loadout scrolls that will perform the spells for players.
While they may be added as player-buyable spells in the future, I
thought it best to implement it this way for now in case the spells need
changed or removed.