Token is no longer allowed in the Arena of Kraanan
People have taken up tokens and went to the Arena of Kraanan so nobody
can get the token unless they log off. This fix will make you drop the
token of you go through the door into the arena (not the foyer) just
like you drop a token when you enter a guildhall.