Nexus node only requires one token
This is a solution to people not having to collect all the tokens to get
the Nexus node. It will be ALOT easier to get that node but already most
players already have it, so it is just to help new players to get it
aswell and not having to include admins to assist us everytime.