Update dement.kod
There was an pullrequest earlier in the year (i guess it was Taran) who lowered the durration from 10-30 minutes. After rewrote close to everything in dement and it seamed to work with hardcoded numbers i checked the original dement.kod again on my testserver and it workes like it should work. Im sure that the problem is the source you compiled from for the live servers and you simple forget to add the change taran made for the spell.