Mules are not allowed in boss zones
Players are abusing mules to make boss zones much easier than they were
designed to be, one example:

Player X wants to do chest runs in trashers, he places his mules in each
"trap room" so the doors will never close, he can then easily run
through and keep doing just the chest without killing all the trashers
on the way.

Boss zones is supposed to be hard and not easy to do solo as it is now.

If a player needs a mule to carry items they can just place them outside
the side, so for trashers they can just place them within the wall
outside the door and just run in there and handover the loot they
collected.