CV Roomflag back to normal
The reason is simple, ppl over 100 HPs are camping there to get easy and safe imps and of course with a armee of mules and of course unattackable. The game is about pvp and CV is a big part of balance it. The dmg is capped by 33% of the full HPs. So it doenst matter if someone has 30 or 99 HPs. If there you want to protect the noobs set the guardianangle up to 100 HPs.