CV Roomflag back to normal
Ppl over 100 HPs are camping there to get easy and safe imps and of course with a armee of mules and of course unattackable. The game is about pvp and CV is a big part of balance it. The dmg someone can make on a other toon is capped by 33% of the others full HPs. So it doenst matter if someone has 30 or 99 HPs. If you want to protect the noobs set the guardianangle up to 100 HPs, but even then the player would only take 4 Hits.

GY is fine and Crypt is fine - but CV is also a premium spot and complete unlogical to the rest of the game.