Revert: invisibility nerf
Invisibility was one of the mayorspells Qor had as a team buff. Qor has also spell names "fade" which works as a focus spell and as a selftarget. The revert has more todo with the following scenario: If you camp a area durring pvp, you need to be invis or a safespot or the mobs will hitting you. Otherwise ppl could say, hey use a invis ring, but the duration of the rings are unaccaptable for camping a area. So back to a spell which was perfect balanced back in the days.