Opt-in Combat Time
Records the last time a player engaged in opt-in combat, which includes
soldier shields, guild wars, and outlaws/murderers.

This will be very useful for future mechanics. This will give us the
capability to differentiate between opt-in combat that players chose to
be part of, and opt-out combat involving PKs attacking unwilling
victims.