safety on / off direct commands with notifying sounds
This change creates two new commands for the client: "safetyon" and
"safetyoff"
These can be bound to hotkeys aliases or typed.

The effect is obvious.  "safetyon" turns the player's safety on while
"safetyoff" turns it off.

I also added a subtle safetyon message sound "click" and a more
aggressive safetyoff noise.

The reason for this change is that safety management is tedious during
pvp.  Players now have access to multiple characters and can
inadventently turn their innocent characters into outlaws or murderers.
The notification sounds provide instant feedback upon logging in as to
the state of the safety and the aliases allow players to quickly disable
or enable the safety without having to fight the interface.