Change appeal user message to include history
Current when a player sends a message via appeal it simply shows the
following:
"If you do not get a timely response, please send game mail to Help for
further assistance."

This makes keeping track of a conversation with an admin difficult.

I changed the appeal message to be more like the tell, say, and other
communication messages.

This is what is seen in the player's message history after an appeal:
"You appeal: [sent message]"