Who list changed to let players fight/talk/use main client
Here is an example video with this change:
http://www.youtube.com/watch?v=pFFCOJCRUuE&feature=youtu.be

The who list window currently functions as a modal dialog box.  This
means that when it is activated it does not allow any input to the main
meridian window.

This change makes the who list window a modeless dialog box, allowing
the player to keep it up while they use the main screen.

One of the main reasons players use multiple clients is to watch the who
list.  This change pushes m59 a little more back towards the one client
per player model.  Everyone watches the who list for pkers and enemies
and since it is not possible to use a client with a who list up, players
run 2 clients.  Accounts are free so there's no reason not to use 2
clients to see the who list.

This change fixes that: now you can watch the who list and play the game
at the same time.