Changed who list window to only open once
I created a function as requested to see if the who list window is
already open when executing the who list command.  It checks if a window
exists with the who list window handle.  If it does not exist then it
creates the who list window and saves the who list window handle.  I
tested it and it works as intended on my test environment.

If it does already exists then the who list window is brought to the
front and the focus is set on it.