Moves the focus to the main window if player has "pain" effect
This change is to help players react in case they are hit and a window
(like the who list) is up that prevents them from taking action
immediately.  Specifically this is a commit tied to the who list dialog
box because the who list blocks the F keys which can allow immediate
logoff.