Lupogg King Lair Fix
Problems with the generator locations made this room very broken. There
are only three generators, one of which was placed in a wall, thereby
preventing spawns. One of the generators is and remains in a tightly
confined space, preventing spawns after a short time.

This led the third generator actually *outside* the Lupogg King Lair to
spawn continually in a single hallway with a safespot right next to it.
It has thus become one of the most abused spots in the game.

This commit simply moves the generators. The one in the wall is now in
the tight tunnel, adding to the cluster of dangerous lupoggs one must
pass through, and the abused generator point is now in the lupogg king
chamber itself.