Makes mummies vulnerable to fire.
Goes hand in hand with giving the shadow mummy a chance to drop a firewand to illustrate the concept of resistance weaknesses. -60 is the smallest weakness that triggers an anti immunity message. (Convulses and seems to suffer horribly.) Players already see the stagger message due to the mummies general -20 weakness to weapons.