Added a sound to the caster if his spell is evaded. Added auto evade below minimum threshold duration.
The caster of a blind or dazzle now gets a fizzle sound if his spell is resisted completely.

There is now a threshold value for a minimum duration blind. If the blind is shorter than that, it will be completely resisted and the caster gets the evade message and fizzle sound.