Changed resist mechanic.
Eagle eyes now checks for the duration of the blind after having substracted up to 2.5 seconds from the duration. If this duration is shorter than 2 to 5 seconds (depending on the power of your EE), the spell is totally resisted, and the caster of the spell gets a fizzle sound on top of the evade message. There is no random component to resists anymore. If we want to, we can use iFactor to create different thresholds for dazzle and blind in the future.