Fixed Deflect bugs
Deflect was calling the attacker's deflect state, when in fact it is the
victim that has the deflect enchantment. This resulted in a 'chance to
deflect' of zero or nil.

Deflect also had its success check backwards, so that a zero or nil
value always succeeded. This was causing deflect to reflect 100% of all
attack spells.

With both of these fixed, the spell now reflects 10% - 75% of all attack
spells based on spellpower as intended.

"Your deflect enchantment attempts to bounce Gar's spell back, but is
not strong enough."