made hospice self cast only and instant cast.
This distinguishes recuperate (formerly hospice) as a self heal used in combats and puts it in line with Vampiric Drain when it comes to ease of use. M59 does not have a "target last target" feature, so targeting yourself in order to heal, then getting on with what you were doing is extremely annoying, especially in larger fights. This allows players to heal with the new Recuperate, while sticking to their target, while still allowing the healing of others with minor heal and major heal. 

While some may consider the retargeting hurdle balance, I argue that poor UI functionality should never be used as a balancing tool.