Balanced out chance to hit.
This is how it works:
chance=(Offense/Defense)*500 bound between 125(=500/4) and 2000(=500*4)
if chance > 1000 and chance > random(1001,2000) => double damage
if chance > 1000 and chance < random(1001,2000) => normal hit
if chance < 1000 and chance > random(1,1000) => normal hit
if chance < 1000 and chance < random(1,1000) => miss

This means: At equal offense and defense, you have a 50% chance to hit. The rest of the hit spectrum is perfectly balanced around this value. 

If your defense is 4 times higher, you will only be hit 12.5% of the time, which means a reduction of damage by the factor 4. 

If your offense is 4 times higher, you will do double damage every time, resulting in an increase of damage by the factor 4.

If you do double damage, you will receive a message in black italic, informing you of your crit. Likewise, if you take double damage, you receive a message in red italic. This corresponds to the color coding for damage numbers and avoids confusion of who is taking double damage.

You are most likely to only encounter crits in PvE or against a pathetic buffless caster as a PF.

Test it out on 104, give feedback!