Adjusts offense and defense rating for players...
...giving agility and aim considerably more weight, and reducing the influence of buffs and items.
In short, rating is now capped at 3000.
1000 from agility/aim
1000 from weapon skills
300 from HPs
200 base line to avoid crazy swings in rating at early levels
500ish from buffs and items.

I didn't change the values of buffs and items, but since the rating is now capped at 3 times the old value and the other contributers have been scaled up by about the same factor, that makes buffs and items only one third as effective as before. The big winner are agility and aim. Yay, for stat diversity! 

Please test on 104 together with the changes to chance to hit and monster offense&defense! Thanks!