Added Death Rift Spell
A level 6 Qor spell sold by Zuxana, intended to replace Shadow Rift.

Death Rift is a mobility spell that fits in deliciously with Qor's lore
and character. A player casting Death Rift in a safe area will be
transported to the Underworld, where he may use any of the typical town
portals. If he has cast Death Rift in an unsafe location before, that
location will be bound to his character, and the random Hell Portal will
show the name of his bound location and send him there instead of to a
random inn.

In this way, we eliminate the mule-based prism aspect of Shadow Rift, we
eliminate the problem of portal longevity (Death Rift is one cast per
one use, heavily disincentivizing using it for money runs), and we also
make the spell more widely available to non-rich players. At Qor 6, with
this design, I believe it'll be used far more for PvP purposes than
farming exploits.

Finally, Death Rift warns players in the target location, so PKs will
not simply appear out of thin air if they use this spell. The warning is
a bold text message and a portal sound centered on the bound location.
The sound is a little quieter than a normal spell cast, and won't be
heard if it's far enough away from the player, but they'll always
receive the text message.

Angeled characters may not cast this spell. This spell follows
CanHavePlayerPortal restrictions preventing its use in major puzzles or
guildhalls, and cannot be used in newbie or guest areas.