Fixed nil multiplication/no application of crit damage.
I orginally multiplied damage with the crit factor *after* sending it to AssessDamage. This resulted in the extra damage not actually being applied to the victim and (since AssessDamage hands out nil values on killing blows) multiplication with nil. This fixes both issues by multiplying with the crit factor before damage is sent to AssessDamage.