Autoloot function
When treasure is generated, a UserGet is called for any dropped number
items, saving the player the hassle of constantly looting. It's a normal
Get attempt, so if the user can't hold the treasure or is too far away,
it will fail.

This does not generate a message on success (as that would create large
amounts of spam), but UserGet does notify the user upon failure due to
range, capacity limits, and so on.