Resist Rings improvement
Resist rings simply didn't work, for a variety of reasons. I attempted
to answer those issues here.

* Resist rings no longer drain charges upon equip.

* Resist rings no longer lose charges over time.

* Instead, resist rings lose a charge whenever they resist an attack of
the appropriate element. For example, if you get hit by a touch of
flame, a Resist Fire ring will lose one charge. A Resist Cold ring will
not.

* Messages are given to inform both the player and the attacker that the
ring is working. This will help the player know he's got the right ring
on, and the attacker know that he needs to switch elements. The
resulting combat behavior should be very interesting.

* Ring charges have been reduced to 8-12. Now that ring charges work
more directly, this means a player can count on a ring for defense
against 8-12 spell attacks. That should be good for about one battle;
long enough to make a huge difference, but short enough for an attacker
to eventually wear down your ring.

* Resist rings' descriptions now list their remaining charges.

* Resist rings reveal themselves when worn, skipping the need for a mule
to Reveal them. This is a design issue, but also a common sense one: if
the ring does magic when I put it on, why can't I just try it on and see
what it does?

* Resist rings no longer unequip upon logoff.