Guild Message of the Day
Rank 4 and 5 guild members may set a Guild Message of the Day that all
members logging in will see. The message includes the rank and name of
the person who set it.

'set motd message' sets the Guild MOTD.
'clear motd' eliminates it.

These commands only work on the guild channel (tg). When setting the
MOTD, guild members will not see the guild send itself (it is not sent
out because it is a command).