No halo for outlaws in guild_pk_only zones
If a guild_pk_only zone prevents the attack of a character you could
otherwise hit, that player will not have a red halo. That same player
will still have a red halo elsewhere, such as inns.

This should help clear up some confusion (namely seeing a red halo for
an outlaw in the city and going to attack them, only to find that you
can't). But we keep red halos for enemies in inns for informational
purposes.