Mark of Dishonor can be casted on monsters
A level four spell that cannot be improved by casting it on monsters is
a bit to hard, it can now be casted on monsters.
If casted it on a monster who aggro a player the monster will stop its
aggro towards that player. It buffs the monster so cannot re-cast it on
the mob till the timer goes away.