Spider Queen and Lair Changes
Spider Queen now has a 50% chance when HitSideEffect is called to cast
spider web on you! Since this bitch is lazy and immobile and she is the
hardest SPIDER in the game, she would generally be able to spit a spider
web or two at you....

The Lair has been moved to 80% spiders 20% baby spiders, you're in a
boss room and we had 80% baby spiders.. Srsly?
Also the spawn has been increased by about 10x the speed & double the
number.