Spellbane converted to RE
In an attempt to create an interesting spell, Spellbane now stops all
magic casting by you, guildmates, and enemies in an extremely small
radius (2, equal to scimitar range). Spellpower can boost range to 3 at
spellpower 50 (mystic range) and range 4 at 99 sp exactly (bof range).

But of course, 99 spellpower for Jala in combat is very very difficult
to achieve. Be careful with this song. It may not work out for you the
way that you want :)