Skill & Spell Softcap Settings
piSkillSoftcapPenalty and piSpellSoftcapPenalty are the divisors for the
softcap penalty when a player has more %s in a spell than double his
primary stat.

Currently, the spell penalty is / 4, and the skill penalty is / 5.

Setting either of these to 1 will effectively eliminate the softcap
penalty. Lower numbers are better. A spell softcap penalty of 2 means
imp chances are only halved, rather than quartered.

Values less than 1 will be bound to 1.