Lost Soldier Shield Graphics Fix
The bug was as follows:
- Delete was called in soldier shield first. At the end of delete, it
called NewUnused on itself.

- NewUnused removes from back and removes stats, then propagates.

- NewUnused in defmod.kod removes the graphic, then propagates.

- NewUnused in item.kod calls UndoPlayerArt on itself

- UndoPlayerArt called RefreshPlayerVisualGear, my function for showing
and layering all gear appropriately.

- RefreshPlayerVisualGear slings the soldier shield back on the player's
back and puts the overlay back in.

- Everything returns, and Delete finally deletes the soldier shield.
Unfortunately, the overlay was still there!

To fix this, I changed RefreshPlayerVisualGear's call so that it will
Post rather than Send. That way, gear will be refreshed AFTER everything
returns, and the soldier shield will no longer exist at that time.