Resistance Buffs & Rings balancing
To the delight of mages and the cries of the masses, this pull request
severely nerfs resistance buffs. There are two reasons for this:

1 - New systems of advancement are coming that will allow players to
improve their characters and their gear in terms of resistances. That'd
all be pointless with these buffs still providing almost total immunity
so easily.

2 - The resistance values granted by these spells and rings has always
been rather absurd. A resist buff and a ring can make you completely
immune to an element. Why not just cast all the buffs and switch rings?
Spell attacks are worse than useless.

So, with those in mind, I've tinkered with the resistance values
granted, and a few other things.

* Resist rings now grant 20% vs an element rather than 50%.

* All resist buffs are now self-cast only. Potions remain unchanged.
There are no potions for resist holy/unholy.

* Resist Magic is now 5% to 25% resist magic, based on spellpower.

* Resist Fire, Cold, Shock, and Acid are now all 5% to 25% resist magic,
based on spellpower. They are inherently stronger, though, in that they
are lower level, of different schools, and stack with rings and other
gear resistances.

* Holy and Unholy Resolve now grant between 9% and 20% resistance to
their uncommon elements. They are weaker than the other resist buffs,
but now only cost a Fairy Wing instead of a Fairy Wing and Solagh. Most
of all, they can no longer be Purged. They are both strong enough to
exactly cover the holy/unholy weakness of ivies and orc shields. (20%)