Fixes battle message color for morph.
The color coding was a bit weird when one of the combatants was morphed. This changes the messages to no longer care about your morphed status, but only that of your enemy. If your enemy is a monster or a morphed player, you get blue text. If your enemy is a nonmorphed player, you get the standard PvP text.