Autoloot cleanup & Money streamlining
* Autoloot no longer spams the player with message failures. Instead,
they are given a summary of range failure, capacity failure, and so on.
UserGet now returns a number indicating the reason for failure, and
MsgSendUser now has the option to report=false and not send the message
to the player.

* Reduced bloated cash values for mob kills across the board, hitting
easily firewalled or cheesed mobs especially hard. However, avar
chieftains and avar shamans were not reduced, as their values were
already fair. The new code for cash creation allows for negative
minimums, which means sometimes, the monsters will drop no money. The
proportion of no-money drops is equivalent to the size of the minimum
value compared to the maximum. One mob can drop sums as high as 4k, but
they only drop money 20% of the time, for example. It's good to mix it
up rather than just having every mob dropping a constant amount of
shillings every kill.

* Boss mobs are worth significantly more cash. The Ghost drops 1k - 4k,
the Orc Pit Boss drops 1k - 2k, and the Lupogg King drops 20k - 50k.
Nope, that's not a typo: he only spawns once every 20 minutes, and he is
a very difficult fight, so his rewards should be high.